packages/backend/requirements.txt
# Core backend dependencies (existing agents)
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
google-genai>=1.0.0
tenacity>=8.0.0
PyGithub>=2.0.0
rich>=13.0.0

# FastAPI server layer
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sse-starlette>=1.8.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
httpx>=0.27.0

# FAISS RAG memory
faiss-cpu>=1.8.0
numpy>=1.26.0
sentence-transformers>=2.7.0

# PR-Agent tools
tomli>=2.0.0
Jinja2>=3.1.0
PyYAML>=6.0.0
